+++ src/test/java/com/jisungin/application/review/ReviewServiceTest.java
@@ -4,8 +4,8 @@ import com.jisungin.ServiceTestSupport;
 import com.jisungin.application.review.request.ReviewCreateServiceRequest;
 import com.jisungin.domain.book.Book;
 import com.jisungin.domain.book.repository.BookRepository;
-import com.jisungin.domain.oauth.OauthId;
-import com.jisungin.domain.oauth.OauthType;
+import com.jisungin.domain.user.OauthId;
+import com.jisungin.domain.user.OauthType;
 import com.jisungin.domain.review.Review;
 import com.jisungin.domain.review.repository.ReviewRepository;
 import com.jisungin.domain.user.User;
