+++ src/test/java/com/jisungin/application/reviewlike/ReviewLikeServiceTest.java
@@ -3,8 +3,8 @@ package com.jisungin.application.reviewlike;
 import com.jisungin.ServiceTestSupport;
 import com.jisungin.domain.book.Book;
 import com.jisungin.domain.book.repository.BookRepository;
-import com.jisungin.domain.oauth.OauthId;
-import com.jisungin.domain.oauth.OauthType;
+import com.jisungin.domain.user.OauthId;
+import com.jisungin.domain.user.OauthType;
 import com.jisungin.domain.review.Review;
 import com.jisungin.domain.review.repository.ReviewRepository;
 import com.jisungin.domain.reviewlike.ReviewLike;
