+++ src/test/java/com/jisungin/domain/book/repository/BookRepositoryTest.java
@@ -9,8 +9,8 @@ import com.jisungin.application.book.response.SimpleBookResponse;
 import com.jisungin.domain.book.Book;
 import com.jisungin.domain.comment.Comment;
 import com.jisungin.domain.comment.repository.CommentRepository;
-import com.jisungin.domain.oauth.OauthId;
-import com.jisungin.domain.oauth.OauthType;
+import com.jisungin.domain.user.OauthId;
+import com.jisungin.domain.user.OauthType;
 import com.jisungin.domain.talkroom.TalkRoom;
 import com.jisungin.domain.talkroom.repository.TalkRoomRepository;
 import com.jisungin.domain.user.User;
