+++ src/test/java/com/jisungin/domain/userlibrary/repository/UserLibraryRepositoryImplTest.java
@@ -11,8 +11,8 @@ import com.jisungin.domain.rating.repository.RatingRepository;
 import com.jisungin.domain.review.Review;
 import com.jisungin.domain.review.repository.ReviewRepository;
 import com.jisungin.domain.userlibrary.UserLibrary;
-import com.jisungin.domain.oauth.OauthId;
-import com.jisungin.domain.oauth.OauthType;
+import com.jisungin.domain.user.OauthId;
+import com.jisungin.domain.user.OauthType;
 import com.jisungin.domain.user.User;
 import com.jisungin.domain.user.repository.UserRepository;
 import org.junit.jupiter.api.AfterEach;
